requirements.txt
mysql-connector-python>=8.0.26
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.9.3 